refactor verifier into a library that can be used in multiple places
