+++ istioctl/pkg/install/verifier.go
@@ -1,0 +1,386 @@
+// Copyright Istio Authors.
+//
+// Licensed under the Apache License, Version 2.0 (the "License");
+// you may not use this file except in compliance with the License.
+// You may obtain a copy of the License at
+//
+//     http://www.apache.org/licenses/LICENSE-2.0
+//
+// Unless required by applicable law or agreed to in writing, software
+// distributed under the License is distributed on an "AS IS" BASIS,
+// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+// See the License for the specific language governing permissions and
+// limitations under the License.
+
+package install
+
+import (
+	"context"
+	"errors"
+	"fmt"
+	"strings"
+
+	appsv1 "k8s.io/api/apps/v1"
+	v1batch "k8s.io/api/batch/v1"
+	meta_v1 "k8s.io/apimachinery/pkg/apis/meta/v1"
+	"k8s.io/apimachinery/pkg/apis/meta/v1/unstructured"
+	"k8s.io/apimachinery/pkg/runtime"
+	"k8s.io/cli-runtime/pkg/genericclioptions"
+	"k8s.io/cli-runtime/pkg/resource"
+	"k8s.io/client-go/dynamic"
+	"k8s.io/client-go/kubernetes/scheme"
+
+	"istio.io/istio/istioctl/pkg/clioptions"
+	operator_istio "istio.io/istio/operator/pkg/apis/istio"
+	"istio.io/istio/operator/pkg/apis/istio/v1alpha1"
+	"istio.io/istio/operator/pkg/controlplane"
+	"istio.io/istio/operator/pkg/translate"
+	"istio.io/istio/operator/pkg/util"
+	"istio.io/istio/pkg/config/schema"
+	"istio.io/pkg/log"
+)
+
+var (
+	scope = log.RegisterScope("postinstall", "postinstall verification logger", 0)
+)
+
+// Verifier runs after installation and is run after installation.
+// As the name says, it verifies if installation is successful.
+type Verifier interface {
+	Verify() error
+}
+
+// StatusBasedVerifier checks status of certain resources like deployment,
+// jobs and also verifies count of certain resource types
+type StatusBasedVerifier struct {
+	IstioNamespace   string
+	RestClientGetter genericclioptions.RESTClientGetter
+	ManifestsPath    string
+	Opts             clioptions.ControlPlaneOptions
+	FOpts            resource.FilenameOptions
+}
+
+// Verify implements Verifier interface. Here we check status of deployment
+// and jobs, count various resources for verification.
+func (v *StatusBasedVerifier) Verify() error {
+	if len(v.FOpts.Filenames) == 0 {
+		return v.verifyInstallIOPrevision()
+	}
+	return v.verifyInstall()
+}
+
+func (v *StatusBasedVerifier) verifyInstallIOPrevision() error {
+	iop, err := operatorFromCluster(v.IstioNamespace, v.Opts.Revision, v.RestClientGetter)
+	if err != nil {
+		return fmt.Errorf("could not load IstioOperator from cluster: %v.  Use --filename", err)
+	}
+	if v.ManifestsPath != "" {
+		iop.Spec.InstallPackagePath = v.ManifestsPath
+	}
+	crdCount, istioDeploymentCount, err := verifyPostInstallIstioOperator(
+		v.IstioNamespace,
+		iop,
+		fmt.Sprintf("in cluster operator %s", iop.GetName()),
+		v.RestClientGetter)
+	return show(crdCount, istioDeploymentCount, err)
+}
+
+func (v *StatusBasedVerifier) verifyInstall() error {
+	// This is not a pre-check.  Check that the supplied resources exist in the cluster
+	r := resource.NewBuilder(v.RestClientGetter).
+		Unstructured().
+		FilenameParam(false, &v.FOpts).
+		Flatten().
+		Do()
+	if r.Err() != nil {
+		return r.Err()
+	}
+	visitor := genericclioptions.ResourceFinderForResult(r).Do()
+	crdCount, istioDeploymentCount, err := verifyPostInstall(
+		v.IstioNamespace,
+		visitor,
+		strings.Join(v.FOpts.Filenames, ","),
+		v.RestClientGetter)
+	return show(crdCount, istioDeploymentCount, err)
+}
+
+// nolint: lll
+func verifyPostInstallIstioOperator(istioNamespaceFlag string, iop *v1alpha1.IstioOperator, filename string, restClientGetter genericclioptions.RESTClientGetter) (int, int, error) {
+	// Generate the manifest this IstioOperator will make
+	t := translate.NewTranslator()
+
+	cp, err := controlplane.NewIstioControlPlane(iop.Spec, t)
+	if err != nil {
+		return 0, 0, err
+	}
+	if err := cp.Run(); err != nil {
+		return 0, 0, err
+	}
+
+	manifests, errs := cp.RenderManifest()
+	if errs != nil {
+		return 0, 0, errs.ToError()
+	}
+
+	builder := resource.NewBuilder(restClientGetter).Unstructured()
+	for cat, manifest := range manifests {
+		for i, manitem := range manifest {
+			reader := strings.NewReader(manitem)
+			pseudoFilename := fmt.Sprintf("%s:%d generated from %s", cat, i, filename)
+			builder = builder.Stream(reader, pseudoFilename)
+		}
+	}
+	r := builder.Flatten().Do()
+	if r.Err() != nil {
+		return 0, 0, r.Err()
+	}
+	visitor := genericclioptions.ResourceFinderForResult(r).Do()
+	// Indirectly RECURSE back into verifyPostInstall with the manifest we just generated
+	generatedCrds, generatedDeployments, err := verifyPostInstall(
+		istioNamespaceFlag,
+		visitor,
+		fmt.Sprintf("generated from %s", filename),
+		restClientGetter)
+	if err != nil {
+		return generatedCrds, generatedDeployments, err
+	}
+
+	return generatedCrds, generatedDeployments, nil
+}
+
+func verifyPostInstall(istioNamespaceFlag string,
+	visitor resource.Visitor, filename string, restClientGetter genericclioptions.RESTClientGetter) (int, int, error) {
+	crdCount := 0
+	istioDeploymentCount := 0
+	err := visitor.Visit(func(info *resource.Info, err error) error {
+		if err != nil {
+			return err
+		}
+		content, err := runtime.DefaultUnstructuredConverter.ToUnstructured(info.Object)
+		if err != nil {
+			return err
+		}
+		un := &unstructured.Unstructured{Object: content}
+		kind := un.GetKind()
+		name := un.GetName()
+		namespace := un.GetNamespace()
+		kinds := findResourceInSpec(kind)
+		if kinds == "" {
+			kinds = strings.ToLower(kind) + "s"
+		}
+		if namespace == "" {
+			namespace = "default"
+		}
+		switch kind {
+		case "Deployment":
+			deployment := &appsv1.Deployment{}
+			err = info.Client.
+				Get().
+				Resource(kinds).
+				Namespace(namespace).
+				Name(name).
+				VersionedParams(&meta_v1.GetOptions{}, scheme.ParameterCodec).
+				Do(context.TODO()).
+				Into(deployment)
+			if err != nil {
+				return err
+			}
+			err = getDeploymentStatus(deployment, name, filename)
+			if err != nil {
+				return err
+			}
+			if namespace == istioNamespaceFlag && strings.HasPrefix(name, "istio-") {
+				istioDeploymentCount++
+			}
+		case "Job":
+			job := &v1batch.Job{}
+			err = info.Client.
+				Get().
+				Resource(kinds).
+				Namespace(namespace).
+				Name(name).
+				VersionedParams(&meta_v1.GetOptions{}, scheme.ParameterCodec).
+				Do(context.TODO()).
+				Into(job)
+			if err != nil {
+				return err
+			}
+			for _, c := range job.Status.Conditions {
+				if c.Type == v1batch.JobFailed {
+					msg := fmt.Sprintf("Istio installation failed, incomplete or"+
+						" does not match \"%s\" - the required Job %s failed", filename, name)
+					return errors.New(msg)
+				}
+			}
+		case "IstioOperator":
+			// It is not a problem if the cluster does not include the IstioOperator
+			// we are checking.  Instead, verify the cluster has the things the
+			// IstioOperator specifies it should have.
+
+			// IstioOperator isn't part of pkg/config/schema/collections,
+			// usual conversion not available.  Convert unstructured to string
+			// and ask operator code to unmarshal.
+
+			un.SetCreationTimestamp(meta_v1.Time{}) // UnmarshalIstioOperator chokes on these
+			by := util.ToYAML(un)
+			iop, err := operator_istio.UnmarshalIstioOperator(by, true)
+			if err != nil {
+				return err
+			}
+			generatedCrds, generatedDeployments, err := verifyPostInstallIstioOperator(istioNamespaceFlag, iop, filename, restClientGetter)
+			if err != nil {
+				return err
+			}
+			crdCount += generatedCrds
+			istioDeploymentCount += generatedDeployments
+		default:
+			result := info.Client.
+				Get().
+				Resource(kinds).
+				Name(name).
+				Do(context.TODO())
+			if result.Error() != nil {
+				result = info.Client.
+					Get().
+					Resource(kinds).
+					Namespace(namespace).
+					Name(name).
+					Do(context.TODO())
+				if result.Error() != nil {
+					msg := fmt.Sprintf("Istio installation failed, incomplete or"+
+						" does not match \"%s\" - the required %s:%s is not ready due to: %v", filename, kind, name, result.Error())
+					return errors.New(msg)
+				}
+			}
+			if kind == "CustomResourceDefinition" {
+				crdCount++
+			}
+		}
+		scope.Debugf("%s: %s.%s checked successfully", kind, name, namespace)
+		return nil
+	})
+	if err != nil {
+		return crdCount, istioDeploymentCount, err
+	}
+	return crdCount, istioDeploymentCount, nil
+}
+
+// Find an IstioOperator matching revision in the cluster.  The IstioOperators
+// don't have a label for their revision, so we parse them and check .Spec.Revision
+func operatorFromCluster(istioNamespaceFlag string, revision string, restClientGetter genericclioptions.RESTClientGetter) (*v1alpha1.IstioOperator, error) {
+	restConfig, err := restClientGetter.ToRESTConfig()
+	if err != nil {
+		return nil, err
+	}
+	client, err := dynamic.NewForConfig(restConfig)
+	if err != nil {
+		return nil, err
+	}
+	ul, err := client.
+		Resource(istioOperatorGVR).
+		Namespace(istioNamespaceFlag).
+		List(context.TODO(), meta_v1.ListOptions{})
+	if err != nil {
+		return nil, err
+	}
+	for _, un := range ul.Items {
+		un.SetCreationTimestamp(meta_v1.Time{}) // UnmarshalIstioOperator chokes on these
+
+		// UnmarshalIstioOperator fails because managedFields could contain time
+		// and gogo/protobuf/jsonpb(v1.3.1) tries to unmarshal it as struct (the type
+		// meta_v1.Time is really a struct) and fails.
+		un.SetManagedFields([]meta_v1.ManagedFieldsEntry{})
+
+		by := util.ToYAML(un.Object)
+		iop, err := operator_istio.UnmarshalIstioOperator(by, true)
+		if err != nil {
+			return nil, err
+		}
+		if iop.Spec.Revision == revision {
+			return iop, nil
+		}
+	}
+	return nil, fmt.Errorf("control plane revision %q not found", revision)
+}
+
+// Find all IstioOperator in the cluster.
+func allOperatorsInCluster(client dynamic.Interface) ([]*v1alpha1.IstioOperator, error) {
+	ul, err := client.
+		Resource(istioOperatorGVR).
+		List(context.TODO(), meta_v1.ListOptions{})
+	if err != nil {
+		return nil, err
+	}
+	retval := make([]*v1alpha1.IstioOperator, 0)
+	for _, un := range ul.Items {
+		un.SetCreationTimestamp(meta_v1.Time{}) // UnmarshalIstioOperator chokes on these
+		by := util.ToYAML(un.Object)
+		iop, err := operator_istio.UnmarshalIstioOperator(by, true)
+		if err != nil {
+			return nil, err
+		}
+		retval = append(retval, iop)
+	}
+	return retval, nil
+}
+
+func show(crdCount, istioDeploymentCount int, err error) error {
+	if err != nil {
+		return err
+	}
+	scope.Infof("Checked %v custom resource definitions", crdCount)
+	scope.Infof("Checked %v Istio Deployments", istioDeploymentCount)
+	if istioDeploymentCount == 0 {
+		scope.Info("No Istio installation found")
+		return fmt.Errorf("no Istio installation found")
+	}
+	scope.Infof("Istio is installed successfully")
+	return nil
+}
+
+func getDeploymentStatus(deployment *appsv1.Deployment, name, fileName string) error {
+	cond := getDeploymentCondition(deployment.Status, appsv1.DeploymentProgressing)
+	if cond != nil && cond.Reason == "ProgressDeadlineExceeded" {
+		msg := fmt.Sprintf("Istio installation failed, incomplete or does not match \"%s\""+
+			" - deployment %q exceeded its progress deadline", fileName, name)
+		return errors.New(msg)
+	}
+	if deployment.Spec.Replicas != nil && deployment.Status.UpdatedReplicas < *deployment.Spec.Replicas {
+		msg := fmt.Sprintf("Istio installation failed, incomplete or does not match \"%s\""+
+			" - waiting for deployment %q rollout to finish: %d out of %d new replicas have been updated",
+			fileName, name, deployment.Status.UpdatedReplicas, *deployment.Spec.Replicas)
+		return errors.New(msg)
+	}
+	if deployment.Status.Replicas > deployment.Status.UpdatedReplicas {
+		msg := fmt.Sprintf("Istio installation failed, incomplete or does not match \"%s\""+
+			" - waiting for deployment %q rollout to finish: %d old replicas are pending termination",
+			fileName, name, deployment.Status.Replicas-deployment.Status.UpdatedReplicas)
+		return errors.New(msg)
+	}
+	if deployment.Status.AvailableReplicas < deployment.Status.UpdatedReplicas {
+		msg := fmt.Sprintf("Istio installation failed, incomplete or does not match \"%s\""+
+			" - waiting for deployment %q rollout to finish: %d of %d updated replicas are available",
+			fileName, name, deployment.Status.AvailableReplicas, deployment.Status.UpdatedReplicas)
+		return errors.New(msg)
+	}
+	return nil
+}
+
+func getDeploymentCondition(status appsv1.DeploymentStatus, condType appsv1.DeploymentConditionType) *appsv1.DeploymentCondition {
+	for i := range status.Conditions {
+		c := status.Conditions[i]
+		if c.Type == condType {
+			return &c
+		}
+	}
+	return nil
+}
+
+func findResourceInSpec(kind string) string {
+	for _, c := range schema.MustGet().KubeCollections().All() {
+		if c.Resource().Kind() == kind {
+			return c.Resource().Plural()
+		}
+	}
+	return ""
+}
